requirements.txt
# Core Flask dependencies
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==3.1.3

# Web server and deployment
gunicorn==21.2.0
python-dotenv==1.0.0

# Template engine
Jinja2==3.1.6
MarkupSafe==3.0.2

# Security and utilities
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# Data processing (core packages only)
pandas==2.2.3
numpy==2.2.5

# Database
PyMySQL==1.1.0
mysql-connector-python==9.0.0

# Additional utilities
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2025.2
packaging==25.0

# Web development utilities
beautifulsoup4==4.13.4
pillow==11.2.1

# For Python 3.12 compatibility
typing_extensions==4.13.2
zipp==3.21.0
